requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
bsedata==0.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
cssselect==1.1.0
cycler==0.11.0
databases==0.6.1
dateutils==0.6.12
fake-useragent==0.1.11
feedparser==6.0.10
fonttools==4.36.0
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
multidict==6.0.2
multitasking==0.0.11
nsetools==1.0.11
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pandas-datareader==0.10.0
parse==1.19.0
Pillow==9.2.0
plotly==5.10.0
pyaes==1.6.1
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
Pyrogram==2.0.41
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
tenacity==8.0.1
TgCrypto==1.2.3
tqdm==4.64.0
urllib3==1.26.11
uvloop==0.16.0
w3lib==2.0.1
websockets==10.3
yahoo-fin==0.8.9.1
yarl==1.8.1
yfinance==0.1.74
zipp==3.8.1
